Use xterm escapes to send background color
Long ago this started out as an Apple Terminal script, and then I added
support for iTerm without considering whether there was one way that
would work in both. To make this work in ghostty, I had to use the xterm
escape sequence, and was surprised to find that it worked in both
Terminal and iTerm too!

That means we can use hex colors throughout, so host_color's little
perl program now just does that.